fix: registry tests error
